backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
elasticsearch==8.11.1
pycti==6.0.4
pytenable==1.4.15
azure-identity==1.15.0
msgraph-core==1.0.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.9
python-dotenv==1.0.0
httpx==0.26.0
